requirements.txt
python-dotenv
# Flask[async]
fastapi
pydantic
uvicorn
httpx
heyoo
runpod
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
chromedriver-autoinstaller==0.6.3
datefinder==0.7.3
exceptiongroup==1.2.0
h11==0.14.0
idna==3.6
outcome==1.3.0.post0
packaging==23.2
Pillow==10.1.0
platformdirs==4.1.0
PyAutoIt==0.6.5
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
regex==2023.12.25
requests==2.31.0
schedule==1.2.1
selenium==4.16.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.23.2
trio-websocket==0.11.1
urllib3==2.1.0
webdriver-manager==4.0.1
wsproto==1.2.0
xvfbwrapper==0.2.9
